Refactor internal imports to use API Gateways (facades)
Refactored imports across `toolkit/`, `adapters/`, and `cli/` to use the defined
facades for `core.state`, `core.workflow`, and `toolkit`.

Maintained specific deep imports when items were not exposed through facades
(such as `MCPTool` and `ToolCapability`). Reverted `to_mermaid` in `toolkit/exporter.py`
back to its deep import to resolve a runtime circular import.
Verified the absence of `CircularImportError` through mypy and runtime testing.
