+++ src/coreason_manifest/adapters/providers/openai.py
@@ -3,8 +3,8 @@ from typing import Any, override
 from coreason_manifest.adapters.system.config import AdapterConfig
 from coreason_manifest.core.common.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.core.security.compliance import RemediationAction
-from coreason_manifest.core.state.tools import ToolPack
-from coreason_manifest.core.workflow.nodes import AgentNode, CognitiveProfile
+from coreason_manifest.core.state import ToolPack
+from coreason_manifest.core.workflow import AgentNode, CognitiveProfile
 from coreason_manifest.ports.llm_provider import GenerativeAdapter
