+++ src/coreason_manifest/toolkit/__init__.py
@@ -1,6 +1,8 @@
 from coreason_manifest.toolkit.builder import AgentBuilder, NewGraphFlow, NewLinearFlow, PlannerBuilder
 from coreason_manifest.toolkit.diff_engine import SemanticPatchReport, apply_rewind, compare_flows
+from coreason_manifest.toolkit.exporter import render_agent_card
 from coreason_manifest.toolkit.validator import validate_flow
+from coreason_manifest.toolkit.visualizer import export_html_diagram, to_mermaid
 
 __all__ = [
     "AgentBuilder",
@@ -10,5 +12,8 @@ __all__ = [
     "SemanticPatchReport",
     "apply_rewind",
     "compare_flows",
+    "export_html_diagram",
+    "render_agent_card",
+    "to_mermaid",
     "validate_flow",
 ]
